+++ examples/CorePlotGallery/src/plots/FunctionPlot.h
@@ -1,0 +1,8 @@
+#import "PlotItem.h"
+
+@interface FunctionPlot : PlotItem<CPTLegendDelegate> {
+    @private
+    NSMutableSet *dataSources;
+}
+
+@end