+++ framework/CorePlot-CocoaTouch.h
@@ -19,6 +19,7 @@
 #import "CPTDefinitions.h"
 #import "CPTExceptions.h"
 #import "CPTFill.h"
+#import "CPTFunctionDataSource.h"
 #import "CPTGradient.h"
 #import "CPTGraph.h"
 #import "CPTGraphHostingView.h"